+++ code/src/cljs/sixsq/nuvla/ui/clouds/utils.cljs
@@ -123,3 +123,22 @@
         (cond-> (and (= template-type "coe") (not-empty cloud-user)) (assoc-in [:template :cluster-params :cloud-user] cloud-user))
         (cond-> (and (= template-type "coe") coe-manager-install) (assoc-in [:template :cluster-params :coe-manager-install] coe-manager-install))
         (cond-> (and (= template-type "coe") management-credential) (assoc-in [:template :management-credential] management-credential)))))
+
+(defn infra-support-pull?
+  [{:keys [capabilities] :as _infrastructure-service}]
+  (contains? (set capabilities) "NUVLA_JOB_PULL"))
+
+(defn swarm-manager?
+  [{:keys [swarm-enabled swarm-manager] :as _infra-service}]
+  (and (true? swarm-enabled)
+       (or (true? swarm-manager)
+           (nil? swarm-manager))))
+
+(defn swarm-worker?
+  [{:keys [swarm-enabled swarm-manager] :as _infra-service}]
+  (and (true? swarm-enabled)
+       (false? swarm-manager)))
+
+(defn swarm-disabled?
+  [{:keys [swarm-enabled] :as _infra-service}]
+  (false? swarm-enabled))