+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/utils.cljs
@@ -1,4 +1,5 @@
-(ns sixsq.nuvla.ui.deployment-dialog.utils)
+(ns sixsq.nuvla.ui.deployment-dialog.utils
+  (:require [sixsq.nuvla.ui.clouds.utils :as clouds-utils]))
 
 
 (defn kw->str
@@ -67,31 +68,11 @@
               (seq files) (assoc :files files))
       :href (:id current-module))))
 
-
-(defn infra-support-pull?
-  [{:keys [capabilities] :as _infrastructure-service}]
-  (contains? (set capabilities) "NUVLA_JOB_PULL"))
-
-(defn swarm-manager?
-  [{:keys [swarm-enabled swarm-manager] :as _infra-service}]
-  (and (true? swarm-enabled)
-       (or (true? swarm-manager)
-           (nil? swarm-manager))))
-
-(defn swarm-worker?
-  [{:keys [swarm-enabled swarm-manager] :as _infra-service}]
-  (and (true? swarm-enabled)
-       (false? swarm-manager)))
-
-(defn swarm-disabled?
-  [{:keys [swarm-enabled] :as _infra-service}]
-  (false? swarm-enabled))
-
 (defn infra-app-compatible?
   [{app-subtype :subtype app-compatibility :compatibility :as _module}
    {infra-subtype :subtype :as infra-service}]
   (boolean (not (and (= infra-subtype "swarm")
                      (= app-subtype "application")
                      (= app-compatibility "swarm")
-                     (or (swarm-disabled? infra-service)
-                         (swarm-worker? infra-service))))))
+                     (or (clouds-utils/swarm-disabled? infra-service)
+                         (clouds-utils/swarm-worker? infra-service))))))